Factor out separate `Framework.Preprod` module from `.DSL`
