+++ lib/integration/cardano-wallet-integration.cabal
@@ -115,6 +115,7 @@ library framework
     Test.Integration.Framework.LocalCluster.SendFaucetAssets
     Test.Integration.Framework.Logging
     Test.Integration.Framework.PendingInEra
+    Test.Integration.Framework.Preprod
     Test.Integration.Framework.Request
     Test.Integration.Framework.Setup
     Test.Integration.Framework.TestData
